python bindings: Use Diagnostics as function arguments
This improves the readability of the code and fixes one testsuite bug.
I suppose this was because we only stored the ptr to the diagnostic
in the FixIt iterator. On the way the python object of it could be
deleted and the ptr was freed, but the iterator was still using this ptr.
Now the object is still used and only freed after the iterator is not needed
any more.